Ffmpeg without audio
FFmpeg functions correctly, but the audio does not play.

If it does not work, check that in the .json you have the "ffmpeg" part in this way

{
  "name": "ffmpeg",
  "features": ["postproc"]
}

And if you get an error from the postproc library, open the terminal and run `vcpkg install`, and it should install everything correctly.